// src/client/services/service-proxies.ts
export interface News {
    _id: string;
    title: string;
    state: string;
    description: string;
    imageUrls: string[];
    price: number;
    priceType: string;
    newsType: string;
    owner: any;
    shop: any;
    location: any;
    status: string;
    usedHours: number;
    serialNo: string;
    categoryId: any;
    type: any;
    brand: any;
    model: any;
    weight: number;
    year: number;
    weightRange: any;
    rentalPeriod: any;
    followBy: any[];
    isFollowing: boolean;
    createdAt: number;
    counterFollow: number;
    counterView: number;
}
// ...
export interface Brand {
    _id: string;
    slug: string;
    name: string;
    logoImage: string;
    totalNews: number;
    brandCategories: Category[];
    brandModels: Model[];
    createdAt: number;
}

export interface FindBrandsResult {
    data: Brand[];
    before: string;
    after: string;
}

export interface CreateBrandPayload {
    name: string;
    logoImage: string;
}

export interface UpdateBrandPayload {
    operation: UpdateBrandPayloadOperation;
    payload: any;
}

export interface Category {
    _id: string;
    slug: string;
    name: string;
    imageUrl: string;
    totalNews: number;
    categoryBrands: Brand[];
    categoryModels: Model[];
    createdAt: number;
}

export interface FindCategoriesResult {
    data: Category[];
}

export interface CreateCategoryPayload {
    name: string;
    imageUrl: string;
}

export interface UpdateCategoryPayload {
    operation: UpdateCategoryPayloadOperation;
    payload: Payload13;
}

export interface Model {
    _id: string;
    slug: string;
    name: string;
    brand: any;
    category: any;
    totalNews: number;
    createdAt: number;
}

export interface FindModelsResult {
    data: Model[];
    before: string;
    after: string;
}

export interface CreateModelPayload {
    name: string;
    slug: string;
    brand: string;
    model: string;
}

export interface UpdateModelPayload {
    _id: string;
    name: string;
    slug: string;
    brand: string;
    model: string;
}

export interface Province {
    _id: string;
    slug: string;
    name: string;
    totalNews: number;
}